Overall & Specific Discount

+++ Software Project/src/Discount.java
@@ -1,7 +1,7 @@
 
 public interface Discount {
-	public void setDiscount(double dis);
-	public double getDiscount_amount();
+	public void setDiscount(int dis);
+	public int getDiscount_amount();
 	public double applyDiscount();
 
 }
