+++ Software Project/src/Specific.java
@@ -1,12 +1,12 @@
 
 public class Specific implements Discount{
 	//Discount discount;
-	public double discount_amount=0.0;
-	public void setDiscount(double dis)
+	public int discount_amount=0;
+	public void setDiscount(int dis)
 	{
 		discount_amount=dis;
 	}
-	public double getDiscount_amount()
+	public int getDiscount_amount()
 	{
 		return discount_amount;
 	}
